rearchitect loading abstraction
this is a major redesign of how OME-TIFFs are loaded and should be more
robust and extensible then the original approach. OME-TIFFs are now
treated as a list of images, each containing a list of files where the
data is actually stored. This structure should also be flexible enough
to accommodate support for #3.

- adds support for multiple positions: closes #1
- adds supports for OME-TIFFs split across multiple files: closes #5