refactor to build master IFD list prior to loading data
This design has numerous advantages over the previous one. Namely, now
the mapping of TiffDatas -> IFDs is much better handled and tested.
Additionally, that is used to generate a master IFD list that maps
XYZCTP indices -> files -> offsets so this design should support memory
mapping in the future (see #3).

This is still some work to be done to support dimension differences
between positions and to further tighten and simplify the code.